Job Candidate Review API System/requirements.txt
django==3.0
djangorestframework==3.9.2
